Adding two IFC examples with modeling issues
A) RE19-Sandwich-wall-with-windows:

Example of a multi-layered wall situation where the layers are composed of individual walls. The three windows are modeled as follows:

1) window in the exterior wall, wall profile cut in the interior
2) window in the exterior wall, wall opening in the interior
3) window in the interior wall, wall opening in the exterior

B) RE19-Space-dormer:

This case shows the usage of a dormer window family. This, although looking nice in 3D, is not recognized by the space boundary calculation of Revit.